Move queries to it's domain schema files
